Refactored and fixed DraftOrderControllerTest.
